+++ packages/itwin/grouping-mapping-widget/src/components/GroupingMappingContext.tsx
@@ -129,6 +129,7 @@ export const GroupingMappingContext = (props: GroupingMappingContextProps) => {
   const [hiddenGroupsIds, setHiddenGroupsIds] = useState<Set<string>>(new Set());
   const [showGroupColor, setShowGroupColor] = useState<boolean>(false);
   const [propertiesShowGroup, setPropertiesShowGroup] = useState<boolean>(false);
+  const [groupColors, setGroupColors] = useState<Map<string, string>>(new Map());
   const [groups, setGroups] = useState<Group[]>([]);
   const [numberOfVisualizedGroups, setNumberOfVisualizedGroups] = useState(0);
   const [isOverlappedColored, setIsOverlappedColored] = useState<boolean>(false);
@@ -190,6 +191,8 @@ export const GroupingMappingContext = (props: GroupingMappingContextProps) => {
       setIsVisualizationsEnabled,
       overlappedElementsMetadata,
       setOverlappedElementsMetadata,
+      groupColors,
+      setGroupColors,
     }),
     [
       showGroupColor,
@@ -200,6 +203,8 @@ export const GroupingMappingContext = (props: GroupingMappingContextProps) => {
       currentHilitedGroups,
       isVisualizationsEnabled,
       overlappedElementsMetadata,
+      groupColors,
+      setGroupColors,
     ],
   );